backend/requirements.txt
# =========================
# Core Backend
# =========================
fastapi==0.120.0
uvicorn==0.38.0
starlette==0.48.0
anyio==4.11.0
h11==0.16.0
sniffio==1.3.1
click==8.3.0
colorama==0.4.6

# =========================
# Environment & Config
# =========================
python-dotenv==1.1.1

# =========================
# MongoDB
# =========================
pymongo==4.15.3
dnspython==2.8.0

# =========================
# Authentication (JWT)
# =========================
python-jose==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# =========================
# Validation
# =========================
pydantic==2.12.3
pydantic_core==2.41.4
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2
email-validator==2.1.1

# =========================
# PDF Processing
# =========================
PyMuPDF==1.26.5

# =========================
# Gemini / Google AI
# =========================
google-genai==1.46.0
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-python-client==2.185.0
google-api-core==2.27.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
proto-plus==1.26.1
protobuf==5.29.5

# =========================
# HTTP / Networking
# =========================
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
urllib3==2.5.0
charset-normalizer==3.4.4
certifi==2025.10.5
idna==3.11
six==1.17.0
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
uritemplate==4.2.0

# =========================
# Utils
# =========================
tqdm==4.67.1
tenacity==9.1.2
websockets==15.0.1
python-dateutil==2.9.0.post0

argon2-cffi==23.1.0
